Add user management unit tests

+++ user/controllers.py
@@ -2,12 +2,12 @@ from flask_restplus import Resource, reqparse
 from .models import User
 from werkzeug.security import generate_password_hash, check_password_hash
 from config import db
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required, get_jwt_identity, create_access_token
 from .permissions import admin_required
 
 from .api_model import user_api
 
-@user_api.route('/user/<int:user_id>')
+@user_api.route('/<int:user_id>')
 class UserView(Resource):
     @jwt_required
     @admin_required
@@ -31,13 +31,15 @@ class UserView(Resource):
         """delete user by id; only admin has access"""
         if user_id is not None:
             user = User.query.filter_by(id=user_id).first()
+            if user ==None:
+                return {"message": "No such user found."}
             db.session.delete(user)
             db.session.commit()
             return {"message": "User deleted."}
         else:
             return {'message': 'Specify the user.'}
 
-@user_api.route('/user')
+@user_api.route('/list')
 class UserListView(Resource):
     @jwt_required
     @admin_required
